Improve the access to first and last elements in LinkedList
Fixes part of bug 594868.

The signature of first and last properties in List did not take in account the
fact that lists permit null elements. Their type has been changed from G? to G
to reflect this.

Also we optimized the access to first and last elements in LinkedList, which can
be made through head and tail pointers directly.