Add ArrayList.sort_with_data
Add a variant of List.sort that takes CompareDataFunc rather than
CompareFunc so compare func could be a method or closure. We are
adding this to ArrayList rather than List to not break the API/ABI.
In 0.7.x, this method will be removed as List.sort will then do
exactly the same.